chore: restructure multi roadview, photo game components #11
