+++ src/features/game/multiplayer/photo/views/GameView.vue
