+++ src/features/game/multiplayer/roadview/components/playerlist/IndividualPlayerList.vue
