+++ src/features/game/multiplayer/roadview/views/TeamRoadGameView.vue
@@ -122,7 +122,7 @@
 </template>
 
 <script>
-import BaseMultiRoadViewGame from "./BaseMultiRoadViewGame.vue";
+import BaseMultiRoadViewGame from "./BaseGameView.vue";
 import ChatWindow from "@/views/Game/MultiplayerMode/components/gameplay/chat/IndividualChat.vue";
 import TeamGameResults from "@/views/Game/MultiplayerMode/components/gameplay/results/MultiplayerTeamGameResults.vue";
 import TeamPlayersList from "./components/TeamPlayersList.vue";
